refactor: LottoTicketCount에서 구매금액을 계산하도록 변경

+++ src/main/java/lotto/domain/game/LottoRevenueRate.java
@@ -8,7 +8,7 @@ public class LottoRevenueRate {
         this.rate = rate;
     }
 
-    public static LottoRevenueRate of(long revenue, long totalTicketMoney) {
+    public static LottoRevenueRate of(long revenue, int totalTicketMoney) {
         double revenueRate = (double) revenue / totalTicketMoney;
         return new LottoRevenueRate(revenueRate);
     }
